# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "EduJobs Connect API",
    "description": "Comprehensive API for the EduJobs Connect platform - connecting Rwandan students with universities and job opportunities",
    "version": "1.0.0",
    "contact": {
      "name": "EduJobs Connect Support",
      "email": "support@edujobsconnect.rw",
      "url": "https://edujobsconnect.rw"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Development server"
    },
    {
      "url": "https://api.edujobsconnect.rw/api",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Authorization header using the Bearer scheme. Enter your token in the text input below."
      }
    },
    "schemas": {
...
      "University": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "objectId",
            "example": "507f1f77bcf86cd799439011"
          },
          "name": {
            "type": "string",
            "example": "University of Rwanda"
          },
          "description": {
            "type": "string",
            "example": "Leading university in Rwanda offering quality education"
          },
          "type": {
            "type": "string",
            "enum": ["public", "private"],
            "example": "public"
          },
          "established": {
            "type": "integer",
            "example": 1963
          },
          "location": {
            "$ref": "#/components/schemas/UniversityLocation"
          },
          "contact": {
            "$ref": "#/components/schemas/UniversityContact"
          },
          "accreditation": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["HEC", "MINEDUC"]
          },
          "programs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Program"
            }
          },
          "rating": {
            "$ref": "#/components/schemas/UniversityRating"
          },
          "statistics": {
            "$ref": "#/components/schemas/UniversityStatistics"
          },
          "featured": {
            "type": "boolean",
            "default": false,
            "example": true
          },
          "status": {
            "type": "string",
            "enum": ["active", "inactive"],
            "default": "active",
            "example": "active"
          },
          "logo": {
            "type": "string",
            "format": "uri",
            "example": "/uploads/universities/ur_logo.png"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uri"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Validation failed"
          },
          "message": {
            "type": "string",
            "example": "Please check your input data"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "current": {
            "type": "integer",
            "example": 1
          },
          "pages": {
            "type": "integer",
            "example": 5
          },
          "total": {
            "type": "integer",
            "example": 50
          },
          "limit": {
            "type": "integer",
            "example": 10
          }
        }
      }
    }
  },
  "paths": {
...
    "/universities": {
      "get": {
        "tags": ["Universities"],
        "summary": "Get all universities",
        "description": "Retrieve universities with filtering, search, and pagination",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "type": "integer",
              "default": 12,
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Search query for university name or description",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "city",
            "in": "query",
            "description": "Filter by city",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "province",
            "in": "query",
            "description": "Filter by province",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "featured",
            "in": "query",
            "description": "Show only featured universities",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sort order",
            "schema": {
              "type": "string",
              "enum": ["name", "rating", "newest"],
              "default": "name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Universities retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "universities": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/University"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/universities/{id}": {
      "get": {
        "tags": ["Universities"],
        "summary": "Get university by ID",
        "description": "Retrieve detailed information about a specific university",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "University ID",
            "schema": {
              "type": "string",
              "format": "objectId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "University retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "university": {
                      "$ref": "#/components/schemas/University"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "University not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/jobs": {
      "get": {
        "tags": ["Jobs"],
        "summary": "Get all jobs",
        "description": "Retrieve jobs with filtering, search, and pagination",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "type": "integer",
              "default": 12,
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Search query for job title or description",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "in": "query",
            "description": "Filter by employment category",
            "schema": {
              "type": "string",
              "enum": ["technology", "finance", "healthcare", "education", "engineering", "marketing", "sales", "other"]
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "Filter by employment type",
            "schema": {
              "type": "string",
              "enum": ["full-time", "part-time", "contract", "internship", "freelance"]
            }
          },
          {
            "name": "level",
            "in": "query",
            "description": "Filter by employment level",
            "schema": {
              "type": "string",
              "enum": ["entry", "junior", "mid", "senior", "lead", "executive"]
            }
          },
          {
            "name": "location",
            "in": "query",
            "description": "Filter by location",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "city",
            "in": "query",
            "description": "Filter by city",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "province",
            "in": "query",
            "description": "Filter by province",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "remote",
            "in": "query",
            "description": "Show only remote jobs",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "featured",
            "in": "query",
            "description": "Show only featured jobs",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sort order",
            "schema": {
              "type": "string",
              "enum": ["newest", "oldest", "salary", "title"],
              "default": "newest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Jobs retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "jobs": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Job"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}